dom/tests/mochitest/bugs/test_bug548828.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=548828
-->
<head>
  <title>Test for Bug 548828</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=548828">Mozilla Bug 548828</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 548828 **/
  var head = document.getElementsByTagName("head")[0];
  is(document.head, head, "document.head and document.getElementsByTagName(\"head\")[0] should match");
  document.head = "";
  is(document.head, head, "document.head reassigned");
  document.head = document.createElement("head");
  is(document.head, head, "new element created with the name \"head\"");
  document.documentElement.appendChild(document.createElement("head"));
  is(document.head, head, "new child named head appended to document");
  var head2 = document.createElement("head");
  document.documentElement.insertBefore(head2, head);
  is(document.head, head2, "new head inserted before existing head");

</script>
</pre>
</body>
</html>
